refactor: remove archetype folders
