+++ IsekaiMod/Content/Classes/IsekaiProtagonist/Archetypes/EdgeLordArchetype.cs
