+++ IsekaiMod/Content/Classes/IsekaiProtagonist/Archetypes/HeroArchetype.cs
