+++ IsekaiMod/Content/Classes/IsekaiProtagonist/Archetypes/MastermindSpellList.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using TabletopTweaks.Core.Utilities;
 using static IsekaiMod.Main;
 
-namespace IsekaiMod.Content.Classes.IsekaiProtagonist {
+namespace IsekaiMod.Content.Classes.IsekaiProtagonist.Archetypes {
 
     internal class MastermindSpellList {
